refactor: extract CompileLifecycleRecoveryCoordinator from CompileController
Move the delegate-injected recovery coordinator (start-timeout handling,
missed-callback handling, watchdog fault recovery) into its own file so
CompileController.cs drops from 619 to 482 lines. No behavior change; the
watchdog lifecycle, TaskCompletionSource semantics, and MainThreadSwitcher/
ConfigureAwait(false) call sites in TryCompileAsync are untouched, and the
characterization tests added in the previous commit stay green.
